Setup Express server for Heroku
