--- integration/package.json
@@ -1,10 +1,0 @@
-{
-  "dependencies": {
-    "@types/node": "^13.7.0",
-    "python-shell": "^1.0.8"
-  },
-  "devDependencies": {
-    "ts-node": "^8.6.2",
-    "typescript": "^3.7.5"
-  }
-}
