+++ package.json
@@ -1,0 +1,14 @@
+{
+  "scripts": {
+    "start": "pipenv run npx ts-node ./src/example.ts"
+  },
+  "dependencies": {
+    "@types/node": "^13.7.0",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "python-shell": "^1.0.8",
+    "@types/express": "^4.17.2",
+    "ts-node": "^8.6.2",
+    "typescript": "^3.7.5"
+  }
+}
